+++ data-raw/FixIEAData.R
@@ -1,6 +1,9 @@
 #
 # This script stores data for internal use by functions that fix 
-# aspects of IEA extended energy balance data
+# aspects of IEA extended energy balance data. 
+# 
+# When any replacement data change, 
+# this script should be sourced before building the package.
 # 
 
 library(dplyr)
@@ -18,10 +21,11 @@ library(tidyselect)
 # Our approach to this problem is to smooth out the really big peak in PSB consumption 
 # by reducing the per-capita consumption of PSB, starting in 1991.
 # The details of this process are recorded in the file "GHAPSB.xlsx".
-# We read the data here and make it available for use internally to the package.
+# We read the data here and gather it.
+# Then, we make it available for use internally to the package.
 # The function fix_GHA_psb() makes use of these data.
 # FixedGHA_PSB <- read.delim(file = file.path("data-raw", "FixedGHPSB.tsv"), 
-#                          check.names = FALSE, stringsAsFactors = FALSE) %>% 
+#                            check.names = FALSE, stringsAsFactors = FALSE) %>% 
 Fixed_GHA_PSB <- openxlsx::read.xlsx(xlsxFile = file.path("data-raw", "GHAPSB.xlsx"), sheet = "FixedGHPSB") %>% 
   tidyr::gather(key = "Year", value = "E.dot", tidyselect::matches("^-?\\d+$")) %>% 
   dplyr::filter(
@@ -32,9 +36,9 @@ Fixed_GHA_PSB <- openxlsx::read.xlsx(xlsxFile = file.path("data-raw", "GHAPSB.xl
     Year = as.numeric(Year)
   )
 
-
-
-# Use these data frames as internal objects in the package
+# Use these data frames as internal objects in the package.
+# These objects are stored in R/sysdata.rda and can be accessed by
+# IEATools:::Fixed_GHA_PSB, for example.
 usethis::use_data(Fixed_GHA_PSB, internal = TRUE, overwrite = TRUE)